dnsdist: Extract the logic in SetEDNSOptionAction into a separate function
So that we can reuse and test it without linking issues.

(cherry picked from commit 721569c13d64fc17aa4b6fd420da8556f5917d7f)
